Reload spaces re-work
The following re-works the reload spaces by moving it up one level into
the facade level. This allows us to make better descisions about how to
reload the spaces.

The logic that defines how and what reload spaces should do, should not
be at the state (model), but instead be controlled via the controller.

This isn't as simple as one would expect, because others are also
reloading spaces add different times.

It's a shame that we can't put this in core package so others can also
gain this logic.

I'll think about how this maybe possible for the agent.