Enforce quota limits for setting the charm and uniter state data
This commit modifies the SetState and SetStateOperation methods to
accept an extra argument that specifies a limit to be applied when
updating the charm and unit state data.

The reasoning behind having two quotas is that it allows us to block
misbehaving charms from storing state data that exceeds a threshold
while at the same time still allowing the uniter to persist its internal
state which may contain information about failed hooks etc. The
individual quota limits may also be set to zero by the operator to
bypass quota checks alltogether. In this case, the max limit is enforced
by mongo's hard limit of 16M per document.

Apart from checking the total (operator-configurable) size for charm
state data, the setter will additionally check that the keys and values
in the state map satisfy the *fixed* length limits which are described
in quote/fixed_limits.go. The limit values are provided by the quota
package since they also need to be enforced by the uniter worker when
charm authors invoke the 'state-set' hook tool.