Introduce controller config options for managing per-unit state quotas
The two options come with sensible defaults, 2M for charm-related state
data and 512k for the uniter's internal state data. Mongo imposes a hard
limit of 16M per document so the two options combined cannot exceed that
hard limit and will trigger a validation error if the do.

Note that the defaults where selected so that we give enough breathing
room for storing charm data but not large enough so that charm authors
are discouraged from mis-using this feature for storing large
CLOBs/BLOBs.