Move CallContext out of State
The CallContext has no need to be in state, it use to depend on state,
but in reality it doesn't. Supplying an interface that satisfies the
dependency requirements is enough to get what we want.

By removing this, we remove the dependency on state from within the
api server. We adhear to better vertical boundaries and use dependency
isolation to give what we only need. This is a different mindset, but
it's a worthy one, it just requires effort to get there.

This is a long road to do that and biting small chunks off as we go is
a worth while endeavour.