Relocate reload spaces
This is a baby step move to allow all other locations to access reload
spaces.

Eventually with more knowledge, we believe that reload spaces should
actually be on the environ and the results should then be sent to state.
At the moment it's a bit backwards and that's a shame.

In reality ReloadSpaces should return a list of NetworkPlanes (for a
better word). Those can then tell you if it's a space or a subnet,
sending each space to be saved to spaces collection and the subnets to
the subnet collection.