Update presence recorder to handle race during server connection.
During the start sequence of an apiserver, there is a potential race
between the agent connection events, and the event handling in the
presence worker. The pubsub forwarder worker is involved in this race
as well.

When pubsub message forwarding is established between two HA nodes
and event is published. The presence worker responds to this event
by asking the other side to let them know their understanding of
presence.

If an agent connection event is published on the controller after
a request for presence state has turned up, but before the presence
worker has responded, the event sequence on the other controller
will get the agent connection before the presence response. However
the presence worker will create the response before it has processed
the new agent connection. If this occurs, one side thinks that the
agent is there while the other doesn't know because it replaced
all that it knew about the other server with the update.

The fix for this is to have the presence recorder only remove
entries from the recorder that aren't alive. When pubsub forwarding
is disconnected, the worker will mark the other server as down
in its own recorder, which updates their state to missing. If
an agent connection event is received before the presence response
message, then we want to keep the alive connection information and
augment with the response from the other side.