Remove the old apiserver metrics.
When we first had metrics for the apiserver they were incorrectly using
the "api" subsystem. We added new ones that used "apiserver" as the subsystem
but left the old ones around while we transitioned.

The old ones are now removed. Originally we had a summary vec and a counter
not realising that the summary vec also gave a count. This is why we are removing
two and leaving one metric.