Enforce max key/value length limits when charms call state-set
The command docs for state-set have been updated to include the
aforementioned limits. Note that the uniter only enforces the fixed
limits which allow us to give some useful early feedback to the charm
that the charm author is probably attempting to mis-use this particular
feature.

As the changes are buffered in memory and committed in bulk when the
hook context is flushed (after the hook scripts have finished
executing), quota checks for the total charm state size are enforced by
the controller. This makes the implementation simpler as the uniter does
not need to fetch and/or watch the controller's configuration to obtain
the active limits. What's more, the buffer-and-let-the-controller-sort-it-out
approach exhibits better performance charactericts as hooks (either
directly or via the operator framework) may call state-set multiple
times and we certainly want to avoid having to serialize the entire
state for each call to enforce the quota limit.