Remove checks in the upgrader that prevent downgrading
This is needed to support agents rolling back to an earlier Juju
version after restoring a backup taken before an upgrade.

The only downgrade that is prevented now is downgrading to a
1.x (which won't be supported by juju-restore).

Also makes the same change in the caasupgrader.

With these changes restoring a previous version of a controller
automatically triggers the agents to download the required version of
the agent binary, symlink it and update the upgradedToVersion field in
the agent.conf.