Move relation state to controller
State data internal to uniter relations has moved from a local file on
the unit to the controller.

Replace StateOps with a single StateManager.
Instead of having each relation state handle its own reads and write.
Create a cache of relation state to be shared.  This allows for all
relation state data to be persisted as one and a reduction in reads
and writes.