Rework client creation
The PR starts to attempt to correctly model the goose client
abstractions in a way that allows for individual goose clients to have
their own http.Client.

With this step it now makes it possible for nova and neutron to have
their own http.Headers. Previously it wasn't possible to tell nova to
have one set of headers and for neutron to have it's own.

It was changed so we can override the neutron http.Headers independantly
from nova. When only affecting neutron it means our potential risk is
lower as we're not touching nova REST APIs.

---

I've split out the client related stuff into a client.go file. This was
to help grok the changes, but if it makes it easier to review I'm more
than happy to put them back and move them in another commit.

I also introduced a ClientFactory which hopefully makes it much easier
where the clients come from.

---

Also added a few comments on the way through and a few TODOs.