Persist meter status state to the controller when using Connected worker
The meterstatus worker is special in that it consists of two workers: a
connected worker that can talk to the controller (e.g. to watch things
or persist its state) and an isolated worker that runs when no
connection to the controller is available.

As our goal is to (eventually) persist the metric status state to the
controller without losing any data, the isolated worker still uses a
disk-backed state store mechanism. When starting, the connected worker
will first check whether there is any state present on disk (either
because the isolated worker was briefly running OR we upgraded from an
older juju version which always stored the state on disk). If that's the
case, then the worker will persist the found state to the controller and
then delete it from disk.

This step also serves as an auto-magic upgrade step for the worker so we
don't need a dedicated update step for this.