Bump version of the MeterStatus facade
This would not be required under normal circumstances as the agents that
are supposed to be calling the V2 methods (provided via a mixin) are
guaranteed to be upgraded after the controller. Unfortunately, pylibjuju
exposes the agent facades which technically makes it possible for
someone to actually try to invoke the new methods on an old controller.