Rename WriteUniterState API call and use the ModelOperation-based state API
Given that this call has the potential to set the state for the unit
(and other agents once/if they get unified) agent and not just the
uniter worker, it has been renamed to WriteAgentState.

Note that this API was introduced for 2.8 and no facade bump is needed.

In addition, instead of calling SetState on the unit, we now call
SetStateOperation and apply the returned operation.