Remove dead spaces
Attempt to remove dead spaces when performing reload spaces. At the
moment it only attempts to remove the spaces if they're dead. That way
we can ensure that we don't remove something accidently.

Failure of removing a space doesn't return an error, it just logs out a
info message, I'm actually unsure about this. I'm more inclined to
collate all the errors up and pass them up to the ReloadSpaces callsite.

For now this might be better off.