Ensure spaces are dead
The following change is somewhat radical, in that it ensures the death
of a space if it's alive. That way we get what we want by removing the
space.

This isn't the way we want it to work in the future, but for now this
solution works the way we think it makes sense right now (future peole
will need to refer to [1])

Future solutions should take more care around the topology and the
connections between nodes in the graph.

1. https://www.sandimetz.com/blog/2016/1/20/the-wrong-abstraction