Don't remove a space if it's a model constraint
If a model has a constraint that is found within a space name, then
don't remove it. This is super cautious, but will probably save our
bacon as the operator has decided that they're letting juju know that a
space is important to them. Removing that space might be somewhat
strange to them, as they've specified that it be there.

The operator flow might turnout to be...

 - Add space with a bootstrap/deploy/model config
 - Reload spaces
 - Move spaces around

Removing step 1 seems counter productive.