Take model endpoint bindings into consideration
Additionally we also need to carefully look at the model endpoint
bindings per application when we're attempting to remove a space. As
spaces might be in use else where and because we don't have the Life
checks for spaces in the code, we can't guarantee we won't break
something there.

There might be other places we also need to check, but the current run
down is:

 1. Don't delete if you're the alpha space
 2. Don't delete if you're the default space (model config)
 3. Don't delete if you're in a model endpoint binding
 4. Don't delete if you're a contraint of a model

Hopefully these should cover us for now, but we should be more ruthless
when deleting if spaces where to check the life before use.