Relocate logic out of state
This is some mighty work to get this working. Most of the logic was deep
down in state, that's bad. Persistence layer should not perform any
complex logic. The following lifts more of that logic out of state. It's
currently inside environs/space but could easily be lifted one level
higher.

This change isn't yet finished, more commits in this PR to come. We're
doing some really magical stuff in here and using shims as adapters.
This means that changing stuff is almost near impossible without
significant changes.

Baby steps with:
 - Adding tests for state life
 - Update mocks