Move to a ReloadSpacesAPI
Coming from a different tangent, rather than forcing the API to use the
networking common API. Create a new API struct that we can abstract out
and test without conforming to any existing interfaces.

If this is turns out to be a good pattern, we can replicate this else
where for better testing of the API layer. These turn out to be more
like commands from an MVC pattern. They hold business logic to be
executed, before they reach state.