API boundary for ReloadSpaces
The changes here create a boundary around ReloadSpaces, one that allows
you to deal with your abstractions and dependencies as they are vs how
you think they are.

Unforunately it creates a raft of code, each code fragment is small, but
tests only at it's level and never goes down or above what it should.

This has the effect of knowning what the API can do, but it does mean
that Juju needs to grow up about understanding this. No longer can it
reach down into state, twiddle a few knobs and come back. This is
prevented and it shows here.

We're testing one thing and one thing only, reload spaces function call.

Everything else is done else where.