Bump charm.v6 version
The updated pkg version does away with forcing a default limit of 1 for
peer and required relations in charm metadata. This is a required change
for correctly enforcing the max-relation limits (e.g. the default value
would cause subordinate metrics collection charms to break).

A caveat of this bump is that lots of tests that expected the default
limit to appear in the parsed charm metadata broke and had to be fixed.