Add origin to ip addresses
Adding an origin field to all ip address with upgrade steps to always be
provider by default, should allow us to reason about whom owns the ip
address.

Once we know about the owner of the ip address we can then think about
the deletion of the ip address if it's no longer used.

The upgrade steps in this PR ensure that we always have on and that it's
always set to the default origin; Provider.

----

Technically we expect the machiner (worker) to never delete any ip
addresses it doesn't own (origin in this case). So if a machiner sees
that the owner is a "provider" we skip this. If it sees the owner is a
"machine" and the ip address is no longer in use, the machiner can then
delete it from the database.