Attempt to remove default installed mongo
The github action now automatically adds mongo from multiiverse of all
places and then installs it. There is no way to say, "nope, don't want
that!" - instead we have to get out our hammer and treat everything as a
nail.

This is not great, we can't rely on a clean environment anymore.