Improve comments.
Adding information about how we expect that the instancepoller will
quiesece the correct addresses after the first run after an upgrade.

It is expected that it *may* take sometime depending on the Short/Long
poll nature of the instance poller.

The state of the network.Origin is a statemachine as follows:

    OriginProvier -> OriginMachine -> Dead