Convert a machine link layer device to locus
The following consumes a machine to give you the current lay of the land
of the topology of the link-layer and addresses. We should be able to
then look at what is being inputted from the args to work out the
difference.

More work is required around that, but we can construct a locus from a
machine at any point.