Move go vet github actions tests to own run
The issue is that we're running out of swap space when attempting to run
go vet. This seems like a 2.8 (develop) issue over 2.7. Something has
changed in 2.8 that blows the testing/building of juju out of the
water.

We really should investigate this as it's only going to get worse in the
future.