Introduce origin to network interface info
In order to propagate network.Origin throughout we need to propagate the
origin via the api server. To do this correctly means updating all the
call sites to spit out the origin. That way we can then consume the
interface info correctly for futher processing.

This is a somewhat drastic change in that you have to touch all
providers and container sites.

I'm unclear as to what the origin type for the containers should be, but
considering they act as providers in that sense and the machiner is the
one checking for the observeable network interfaces that seemed to make
sense.

I've not put any tests up yet as I'm unsure if this is the way we
actually want to go, but for me this is the cleaner design.