Ensure NetworkOrigin is always set
The following is a bit convoluted as we want to version bump the
MachinerAPI along with NetworkConfigAPI. It turns out that most people
don't bump the NetworkConfigAPI, instead they mark things as omitempty
and just fudge it at a later date.

I believe this is wrong and hides the fact that this is a potential for
API facade inconsistencies. Instead what we do here is bump the
NetworkingConfigAPI and force the user to pick the right API version
based on the behaviour they desire.

I actually believe the composition of these facades aren't such a great
idea for juju. Instead they should have been realised struct members.