Re-use the VPC
During a manual deploy test on aws, VPC's are like gold dust. Getting
more of them is possible, but painful. To that end we should reuse them
for our manual deploy test. That way multiple tests can in theory run at
the same time as it should notice the VPC is up and running.

This is pure cattle set up, if the VPC isn't there it will create a new
one. If it is, it will reuse it.