Only allow spaces in constraint values if they are slash-escaped
A fix that landed previously allowed unescaped constraint values with
spaces but made the constraint-parsing code more convoluted and relied
on the removal of already-set checks to function properly.

This commit reverts the previous set of changes and introduces a better
way for dealing with constraint values with spaces by expecting the
operator to slash-escape them as you would normally do in a shell.