Find datastores in folders
Update the vsphereclient Datastores method to return datastores under
foldes under the Datastores root folder. We don't keep track of the
inventory path of the datastore - when the name of the datastore is used
for creating a new vm it's looked up on the compute resource, so the
inventory path isn't needed.