Introduce ReadLocalApplicationSettings API call
This commit deprecates the use of the ReadSettings (which does not work
for k8s charms as described in LP1876097) API for obtaining the local
application databag contents and instead replaces it with a new call
which works as expected for both k8s and non-k8s charms.

Note that ReadSettings is still used to read the data bag for individual
units and must therefore be preserved.