Lots of cleanup and rework.
Handle the case where the script can't be executed in Runner, as well as
a safety fallback in the hook.sh script.

Add actual tests to the hook.sh script that it can gracefully handle the
case when the file it should run isn't executable.

Improve the log message for debug-code so it doesn't say something about
'debug-hooks'.