Fix a logic error in the database upgrade worker.
If one of the HA secondary controllers fails to process the upgrade steps
for whatever reason, like disk space issues, then the upgrade got stuck
unrecoverably.

This is because the worker assumed that the changes to the upgrade info
record in the database would be happening while it was watching. So if the
agent had been bounced, and it wasn't the primary, it would start watching
for changes to the doc, and time out after ten minutes. However the primary
and the other controller had already finished, and the upgrade info doc
had a status of "finishing".

The fix here is two-fold. First, once the watcher has been started,
reread the upgrade info document and check the status before waiting on
the changes from the watcher. If the status is "db-complete" or "finishing"
then the worker records success and is done. Otherwise it waits for the
watcher.