Fix edge case in manual provisioning.
If the machine that you are trying to manually provision has the
2.8/edge juju snap installed, it introduces a snap service called
"snap.juju.fetch-oci". The code was just checking for "juju" in the
returned service list.

We now look for "jujud-machine" as a service name as any machine
that we sould look to provision over SSH would have a machine agent
set up if it was provisioned.