Ensure that reboot requests are queued correctly
A commit which was part of the 2.8 refactoring work messed up the
implementation of the handleReboot method. While the method would return
the correct error (ErrReboot or ErrRequeAndReboot), the actual reboot
request was never queued due to an early return logic bug.

This commit makes everything work again and includes additional checks
to prevent future regressions.