Ensure kvm images are pulled from the configured image URL
While the image fetch code always queries the correct index url (as
specified by the image-metadata model config) for discovering kvm
images, it did not make use of it when fetching the actual images. As a
result, juju would always fall back to using the built in default
(cloud-images.ubuntu.com).

This commit threads the value of the image URL parameter through the
image fetching code and uses it when constructing the URL to pull the
image from.