Ensure that the goroutine forked in RunHook exits.
TestRunHookDebugAt was causing us to fork a goroutine that was waiting
for the process to exit, which potentially sat around forever (because
no client actually connected to the SSH side of tmux). And then the next
test would start, and there was an active goroutine spinning on the
variable waitClientExit. (and it would patch that variable).

The existing mechanism wasn't quite right, because it simulates a client
connecting and immediately disconnecting, which would cause the client
side to kill the bash process.
However, if we put in an unbuffered channel, then we don't try to kill
the child process until RunHook has returned.

This makes 'go test -race' happy for me.