Fix query tracker tests.
All of the tests that were using the query tracker were subject to races
in other goroutines that may also be hitting the database. Since the
query tracker was loaded globally onto the state object the fix for
this is to initialize the query tracker with the method that we are
interested in tracking.

For the status tests, this is "client.(*Client).FullStatus" as it needs
to match the string that is found in the traceback.

Prior to this fix using stress-ng would always cause the test to fail,
with the fix I'm unable to cause a race failure. Also the wait for model
idle methods are now not needed as that was an attempt to deal with
the above problem of other goroutines hitting the database. Now that we
are filtering them out, we don't need to wait any more.