Increase the max wait time for WaitAdvance.
Under load, or on slow instances, the testing.ShortWait time (50ms)
is not sufficient for the dependent code to get to the required situation
of waiting on the clock.

There is no real impact for changing testing.ShortWait for testing.LongWait.
The WaitAdvance code will still check every 10ms for the appropriate number
of waiters.