Fix almost all the machine agent tests that depend on leases
The raft workers needed some extra steps to get them running: the
cluster (of one machine) needed bootstrapping and the raft transport
needs to know its own address before the raft worker can start. This
is normally published by the peergrouper, but the peergrouper was
failing because mongo isn't running with --replSet in these tests. I
initially tried turning it on and getting the replicaset configured
but for reasons I couldn't work out the config changes wouldn't stick.

Instead we get hold of the hub and publish the test address
directly. (It's likely that the code grabbing the hub is racy - I
couldn't check because of the unsafe pointer conversion error when
running the race checker.)

There's one test remaining that relies on claiming a singular lease in
state before starting the agent - I'll fix that next.