Fix TestModelWorkersRespectSingularResponsibilityFlag
This was relying on being able to get a singular claimer from
state. Unfortunately there's no easy way to get the equivalent claimer
from the lease manager from outside the dependency engine - the
API server gets it as a dependency through the manifolds.

Instead inject a snapshot into the raft store in the same way as we do
when migrating from legacy leases in upgrades (but simpler because it
doesn't need to go through the raft logs to find config and current
term and index).