Remove state lease workers and code that uses them
Some generic features are still used for interacting with
state/raftlease.

The state/lease package is kept because it is used by upgrade code.

Remove the code in globalclockupdater that allows it to take either a
state or a lease manager as input - now it always requires a lease
manager (and a raft).