Don't write errors when pubsub websocket disconnects.
Having the other side of the websocket go away isn't an error.
We now check to see if the error is a close error type, and if
it is, just log at trace. Only log an error if there is some other
error type.

Also had the pubsub handler use a more detailed logger for writing
errors so they can be filtered more easily.