Expose the engine worker restart time as a package variable.
This allows tests to change the restart time so older wall clock
type tests don't have to wait three seconds for workers to restart.

In particular some of the MachineSuite tests are impacted by this as
the api caller worker is restarted when the addresses change. Which is
signalled by another worker.

Without this patch, the TestManageModelRunsCleaner test ran in approx
six seconds, ten seconds, or timed out about one in 12. With this patch
the test is a consistent six seconds, which is still too long, but
at least consistent.