Updating the endpoint bindings to take in a lookup
The following updates MapWithSpaceNames to consume a space infos lookup.
This way we can correctly handle the idea of using the space infos once
and then passing it to the endpoint bindings.

Requesting the space infos isn't exactly light and doing it with in a
loop over N applications isn't exactly the wisest move. We gain
substantial speed up for Status with large applications and complex
spaces and deploying a bundle with space names.