Fix bug that may cause data-loss due to accessing map value by reference
The gatherBridgeAddresses receive a map of "toRemove" addresses which
they extend to include the addresses for a specified bridge interface.
However, maps in Go are passed by reference to their internal state
structure. As a result, if the map needs to grow, a new map will be
allocated but the caller's map reference will still point to the old
backing store thus causing loss of data!

To fix this, this commit refactors the code to collect the addresses
and merge them into the map instead.