Sets the caas running chan to buffered.
We are making this change to help fix some race conditions happening in
the unit tests where by the uniter may not have fully started yet or it
currently isn't sitting on the other end of the channel waiting.

By making it buffered by one we can guarantee that the events is not
missed by the uniter.