Lift endpoint bindings call site
The following lifts the endpoint bindings and spaceIdsToProviderIds
(which needs a new name) up the stack so we can then pass that
information down the stack without requesting multiple times.

The aim of this is to better model the dependencies before attempting to
merge the endpoint bindings and constraints together in the space
topology.

This mechanical change should bring a lot of performance gains for large
deployments of machines, as we're only requesting things once per
machine call or even per API request.