Remove spaceIdsToProviderIds and dependencies.
Removal of the spaceIdsToProviderIds is no longer required, as we can
inline that step in the one location. Doing this means we can drop
another call to state.

This in theory, should make provisioning large arrays of machines much
quicker. We should attempt at some point do benchmarking around the
provisioning of machines and see what gains we end up with.