Refactor leadership pinning API
The following re-works the common leadership pinning API. It removes the
idea that the API is part of a facade and instead defines it's own type,
one that can move independantly of a facade version. This has been an
ongoing discussion for sometime, but has concretely formalised to a
discourse post[1].

The changes are relatively simple, adding new methods for taking a
application name slice, so calling Pin/Unpin doesn't require a machine
to talk to. This should give us a lot of flexibility about when and how
we can call these methods.

1. https://discourse.juju.is/t/common-facade-apis/3081