Remove package level logger from worker/meterstatus.
This also includes the adding of the GetLogger method to the
restricted context. This is to ensure that the logging methods during
hook execution.

A few methods ended up getting config structs as args instead of adding
yet more arguments, and I also cleaned up the Validate testing of the
various connected or isolated configs.