Unping machine applications on destroy
During a destruction of a machine application, we should attempt to
remove any last references to ties for leadership pinning.

The normal destroy flow will attempt to remove the leadership pinnings,
but will fail if there exists pinnings. Using force will allow this to
proceed.

The call for removing pinnings is idempotent, so calling it multiple
times is fine.