Move to warnings instead of errors
The following updates the code to warnings, as the machines have
actually been removed and we're telling the callee that we failed to
remove the machine, which is not the case at all.

Note: in some cases if a application has pinned during series upgrade
and it has been pinned without a timeout, then the leadership will
still prevent another leadership change. The work around for this
case until we provide the ability for the operator to unpin via the
CLI, is to remove the raft logs manually.