Filter out INFAN networks for ec2
Filtering out INFAN networks for ec2 prevents the change of landing on
that subnet when we know that an instance can't start there. Instead it
should use the underlay network and that generally is at the start of
the slice.

The fitlering code is rather rudamental, but is generic enough that it
could be lifted into the core package and allow other providers to use
the same logic. Currently Openstack provider implements it's own
filtering, but could be persuaded to use this one instead.