Correctly handle bindings for placements
The following changes attempt to handle both space constraints and
binding placements when looking for subnets in availability zones.

Previously the code looked only for spaces, but binding placements
although similar aren't spaces, so the code path required to do the
lookup was failing for this.

In doing the changes, I also refactored the code to highlight the
similarities (random selection and filtering) and the differences, the
VPC ID and the lack of VPC ID selection.

Additionally I've added some unit tests to ensure that we correctly test
at least some of the environ code, as the time of writing we have very
little to no real unit tests for individual functionality for pure unit
level testing.