Add metrics around the queue length and age for the multiwatcher worker.
This required passing a clock into the worker, which means all the sources
that were using the worker now also need to pass in a clock.

Adds a number of metrics for monitoring the queue length and age, along
with summaries that give us timings for adding items to the pending deque
and how long things are taking to process.