Integration tests for lxd profiles
The integration tests here are to ensure that we add the correct lxd
profile and that we correctly remove the lxd profile. This just setups
that the profiles are correctly installed and update correctly.

More work is required to ensure via the charm that we error out if the
profile isn't applied.