Assign predictable names to host NICs for LXD container virtual NICs
The NIC naming convention is "machineID-deviceIndex" where machine ID is
the juju machine ID (e.g. 0/lxd/1) with the slashes removed. This change
not only makes it easier to spot such interfaces in the 'ip a' output
and but also will allow juju to enumerate their names when the container
is removed (this information is exposed by the LXD API as part of the
container's state).