Charmhub client
The following is a introduction and re-working of the charmhub, focusing
on the info API endpoint.
The client is rather simple, yet also powerful, as you can
interchange the underlying requester. This should allow the changing
from a simple http.Client, to a bakery one. The way it's set up should
allow the sharing of the bakery client, so we can then share the cookie
jar between stores to prevent logging in every time (in theory).

The tests also include integration tests against the real endpoint, we
do this to ensure that we're not missing any API response issues. These
aren't run by default, to run them you have to pass -tags="integration",
to the go test command.