Adding find methods to charmhub.
The following adds find client methods for hitting the find endpoint.
The find is rather dumb at the moment, as we don't currently pass in any
additional arguments (--narrow for example). Instead we're doing a
purely dumb lookup against some canned responses. So attempting to test
how other queries will work seem rather fruitless, we should though wire
these up in the future.

The code itself is rather boring, it just uses the existing rest client
to query the endpoint and marshalls it correctly based on the content
type.