Add fields to determine status for CAAS models.
In order to determine the display status of an Application in a CAAS
model, we need to know whether or not a workload is expected.
This is determined by whether there is a podSpec set for the
application.

If the application has an Unset status, it is derived from the
units. In the CAAS sense, the also takes into consideration
whether the application for the unit is expecting a workload.