Better error messages when deploying bundles
This brings in the new charm and charmrepo changes around error
messages, when reading and parsing bundle files. The charm repo
highlights and identifies each file that went wrong, so the charmer can
see easily how something isn't quite right.

There is one problem and that is the sad reality of the code though, we
now have 3 different error types that we have to negotiate between...

 - std error
 - gopkg.in/errgo
 - github.com/juju/errors

Considering this is in fact a juju project, we should drop the errgo one
and be done with it. The code path itself has to marshal and unmarshal
the errors from std err -> juju/errors -> errgo -> juju/errors and each one has
different semantics. Not to mention we actually loose the stack trace
from the marshalling and unmarshalling, which is a damn shame.

Fixes: https://bugs.launchpad.net/juju/+bug/1871711