Display message with store channel when upgrading charms from the store
The channel from which the charm was fetched is recorded when the charm
gets deployed and the operator has no means to query this information
when they are trying to upgrade their charm.

To amend this issue, this commit adds an info message that includes the
configured channel when upgrading to a charm that is hosted remotely.